Add arguments to cli to control socket creation
Must specify 'inet' or 'unix' as the first argument (or shortened forms)
and then either a port or file system path, respectively.

To make this easier `SocketServer::Initialize` was made virtual with no
parameters, so the `Inet` derived class now takes address and port
during construction, though it doesn't really need to store them for the
lifetime of the object.