Move IO buffers out of `SocketClient`
Currently buffers seem to be "small" enough to fit on the stack. Since
we always either empty the buffers or error out, we never need to retain
any data across calls to these methods.

This does start to suggest that `SocketClient` might be better off as
just a wrapper for closing fds and functions can move to `SocketServer`
for less overhead on `Logger` pointers and better default function
inlining.